+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [Unrelease]
+
+  - Made `github-token` parameter optional
+
+
 ## [v20200412]
 
   - Leverages `with` keyword to configure the action